Conflicts:
	dmff/admp/pme.py
	dmff/admp/settings.py
	dmff/settings.py

+++ dmff/settings.py
@@ -1,6 +1,6 @@
 from jax.config import config
 
-PRECISION = 'float'  # 'double'
+PRECISION = 'double'  # 'double'
 
 DO_JIT = True
